Here my tdd approach and yagni rule hit the wall.
Until now the searching algorithm wouldn't have needed backtracking.
But with this test we see that we are dealing with a huge n-ary
tree data structure and we need backtracking search to find the
solution.
But figuring it out just from tests, without prior algorithm training
would be a huge effort.